pyproject.toml
[tool.poetry]
name = "Python FastApi Base Template"
version = "2.1.0"
description = "My personal base template for FastAPI projects."
authors = []
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.34"
pydantic = { extras = ["dotenv", "email"], version = "^2.10.5" }
dependency-injector = "^4.43.0"
yoyo-migrations = "^9.0.0"
aiocache = { version = "^0.12.1", extras = ["redis"] }
aiopg = "^1.4.0"
bcrypt = "^4.0.1"
starlette-prometheus = "^0.10.0"
pydantic-yaml = "^1.3.0"
aiohttp = "^3.8.2"
pyotp = "^2.7.0"
httpx = "^0.27.0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-logging = "^0.36b0"
opentelemetry-exporter-otlp = "^1.15.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
pydantic-settings = "^2.3.4"
aio-pika = "^9.4.2"
redis = "^5.0.4"
pre-commit = "^3.7.1"
python-json-logger = "^2.0.7"
aioboto3 = "^13.1.1"
prometheus-fastapi-instrumentator = "^7.0.0"
polyfactory = "^2.17.0"
jsf = "^0.11.2"
apscheduler = "^3.10.4"
aiogram = "^3.17.0"
colorama = "^0.4.6"
beautifulsoup4 = "^4.13.3"

[tool.poetry.group.dev.dependencies]
MarkupSafe = "^2.1.5"
lxml = "^4.7.1"

[tool.poetry.group.formatters.dependencies]
autoflake = "^2.3.1"
isort = "^5.13.2"
black = "^24.4.2"
docformatter = "^1.7.5"
add-trailing-comma = "^3.1.0"

[tool.poetry.group.sasts.dependencies]
itsdangerous = "^2.2.0"
safety = "^3.2.4"
bandit = "^1.7.9"

[tool.poetry.group.linters.dependencies]
flake8 = "^7.1.0"
mypy = {extras = ["reports"], version = "^1.11.0"}
ruff = "^0.5.2"

[tool.poetry.group.tests.dependencies]
pytest-asyncio = "0.21.2"
coverage = "^7.6.0"
pytest-cov = "^5.0.0"
pytest-repeat = "^0.9.3"

[tool.poetry.group.benchmarks.dependencies]
locust = "^2.29.1"

[tool.poetry.group.sugar.dependencies]
pytest-sugar = "^1.0.0"

[tool.ruff]
line-length = 88
lint.select = ["E", "W", "F", "C90", "N", "Q"]
lint.ignore = ["E501", "F401", "F403"]

[tool.poetry.scripts]
lint = "ruff ."
typecheck = "mypy ."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
